[#210]fix : controller 분리 및 커스텀 에러 적용
