+++ server/src/middlewares/ErrorHandler.ts
@@ -1,13 +1,13 @@
 import { NextFunction, Request, Response } from 'express';
-
+import * as _ from 'lodash'; 
 // 에러 미들웨어는 항상 (설령 안 쓰더라도)
 // error~next의 4개 인자를 설정해 주어야 함.
 
-class CustomError extends Error {
-  statusCode : number
-  constructor(message : string, statusCode : number ) {
+export class CustomError extends Error {
+  statusCode : number;
+  constructor(message : string, statusCode : number = 500 ) {
     super(message);
-    this.statusCode = 500;
+    this.statusCode = statusCode;
 
   }
 }
@@ -23,4 +23,14 @@ function errorHandler(
   res.status(error.statusCode).json({ result: 'error', message: error.message });
 }
 
-export { errorHandler };
+function blockEmptyObject (obj : Object) : void {
+  if(_.isEmpty(obj)){
+    throw new CustomError(
+      "body가 비어있거나 header의 Content-Type이 'application/json'인지 확인해주세요",
+      400
+    )
+  }
+
+}
+
+export { errorHandler, blockEmptyObject };
