+++ server/src/services/UserService.ts
@@ -1,28 +1,12 @@
 import {
   userModel,
-  UserModel,
-  UserInfo,
-  UserData,
-  StatusInfoRequired,
+  UserModel
 } from '../db';
+import {UserInfo, UserData, StatusInfoRequired, LoginInfo, LoginResult, UserInfoRequired} from '../types/UserTypes';
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
+import {CustomError} from '../middlewares';
 
-interface LoginInfo {
-  email: string;
-  password: string;
-}
-
-interface LoginResult {
-  token: string;
-  role: string;
-  userStatus: string;
-}
-
-interface UserInfoRequired {
-  email: string;
-  currentPassword: string;
-}
 
 enum UserStatus {
   NORMAL = 'normal',
@@ -43,14 +27,7 @@ class UserService {
       role,
       userStatus,
     } = userInfo;
-    const user = await this.userModel.findByEmail(email);
-
-    // 이메일 중복 확인
-    if (user) {
-      throw new Error(
-        '현재 이 이메일은 사용 중입니다. 다른 이메일을 입력해 주세요.'
-      );
-    }
+    
 
     //비밀번호 해쉬화
     const hashedPassword = await bcrypt.hash(password, 10);
@@ -69,6 +46,7 @@ class UserService {
     return createdNewUser;
   }
 
+  
   // 일반 로그인
   async getUserToken(loginInfo: LoginInfo): Promise<LoginResult> {
     const { email, password } = loginInfo;
@@ -76,8 +54,8 @@ class UserService {
     const user = await this.userModel.findByEmail(email);
 
     if (!user) {
-      throw new Error(
-        '해당 이메일의 가입 내역을 찾을 수 없습니다. 다시 한 번 확인해주세요.'
+      throw new CustomError(
+        '해당 이메일의 가입 내역을 찾을 수 없습니다. 다시 한 번 확인해주세요.',400
       );
     }
 
@@ -85,7 +63,7 @@ class UserService {
     const isPWCorrect = await bcrypt.compare(password, hashedPassword);
 
     if (!isPWCorrect) {
-      throw new Error('비밀번호가 일치하지 않습니다. 다시 한 번 확인해주세요.');
+      throw new CustomError('비밀번호가 일치하지 않습니다. 다시 한 번 확인해주세요.', 400);
     }
 
     const secretKey = process.env.JWT_SECRET_KEY || 'secret-key';
@@ -105,7 +83,7 @@ class UserService {
     const user = await this.userModel.findById(userId);
 
     if (!user) {
-      throw new Error('가입 내역이 없습니다. 다시 한 번 확인해 주세요!.');
+      throw new CustomError('가입 내역이 없습니다. 다시 한 번 확인해 주세요!.', 400);
     }
     return user;
   }
@@ -127,8 +105,8 @@ class UserService {
     );
 
     if (!isPasswordCorrect) {
-      throw new Error(
-        '현재 비밀번호가 일치하지 않습니다. 다시 한 번 확인해 주세요.'
+      throw new CustomError(
+        '현재 비밀번호가 일치하지 않습니다. 다시 한 번 확인해 주세요.', 400
       );
     }
 
@@ -161,8 +139,9 @@ class UserService {
         correctPasswordHash
       );
       if (!isPasswordCorrect) {
-        throw new Error(
-          '비밀번호가 일치하지 않습니다. 다시 한 번 확인해 주세요.'
+        throw new CustomError(
+          '비밀번호가 일치하지 않습니다. 다시 한 번 확인해 주세요.',
+          400
         );
       }
 
@@ -172,6 +151,16 @@ class UserService {
       return false;
     }
   }
+
+  //중복이메일있는지 확인
+  async isSameEmail (email : string):Promise<boolean> {
+    const user = await this.userModel.findByEmail(email);
+    if(!user){
+      return false;
+    } else {
+      return true;
+    }
+ }
 
   async getUsers(): Promise<UserData[]> {
     const users = await this.userModel.findAll();
@@ -192,7 +181,7 @@ class UserService {
   async blockExpiredUser(email: string): Promise<boolean> {
     const user = await this.userModel.findByEmail(email);
     if (!user) {
-      throw new Error('유저를 찾을 수 없습니다.');
+      throw new CustomError('유저를 찾을 수 없습니다.',400);
     }
     const userStatus = user.userStatus;
